refactor: Logo na pasta public
